feat(agent): add user notification module
Add user desktop notification module to agent. This module runs a grpc
service that listens for notifications and writes the messages into a
socket. A systemd user socket activated unit sends the notification to
any logged in user with a graphical session.

This pattern separates responsibilities and avoids setuid in the agent.

Signed-off-by: Manuel Bluhm <manuel@ssrc.tii.ae>